feat(auth): moved the JWT functionality to a separate library for ease of use in other microservices.
