+++ client-processing-ms/src/main/java/ru/t1/nour/microservice/config/WebSecurityConfig.java
@@ -10,21 +10,15 @@ import org.springframework.security.config.annotation.method.configuration.Enabl
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import ru.t1.nour.microservice.service.impl.UserDetailsServiceImpl;
-import ru.t1.nour.microservice.util.AuthEntryPointJwt;
-import ru.t1.nour.microservice.util.AuthTokenFilter;
-import ru.t1.nour.microservice.util.JwtProperties;
-import ru.t1.nour.microservice.util.JwtUtils;
+import ru.t1.nour.security.jwt.AuthEntryPointJwt;
+import ru.t1.nour.security.jwt.AuthTokenFilter;
+import ru.t1.nour.security.jwt.JwtUtils;
 
-import java.security.KeyFactory;
-import java.security.PrivateKey;
-import java.security.spec.PKCS8EncodedKeySpec;
-import java.util.Base64;
 
 @Configuration
 @EnableWebSecurity
