+++ spring-common-jwt-security/src/main/java/ru/t1/nour/security/jwt/AuthEntryPointJwt.java
@@ -1,4 +1,4 @@
-package ru.t1.nour.microservice.util;
+package ru.t1.nour.security.jwt;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.ServletException;
@@ -8,13 +8,11 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
-@Component
 @Slf4j
 public class AuthEntryPointJwt implements AuthenticationEntryPoint {
     @Override
