+++ spring-common-jwt-security/src/main/java/ru/t1/nour/security/jwt/JwtUtils.java
@@ -1,16 +1,13 @@
-package ru.t1.nour.microservice.util;
+package ru.t1.nour.security.jwt;
 
 import io.jsonwebtoken.*;
-import io.jsonwebtoken.JwtParser;
 import io.jsonwebtoken.io.IOException;
-import jakarta.annotation.PostConstruct;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
+import ru.t1.nour.security.jwt.properties.JwtProperties;
+import ru.t1.nour.security.jwt.utils.PublicKeyLocator;
 
 import java.security.KeyFactory;
 import java.security.PrivateKey;
@@ -20,7 +17,6 @@ import java.time.temporal.ChronoUnit;
 import java.util.*;
 import java.util.function.Function;
 
-@Component
 @Slf4j
 public class JwtUtils {
 
@@ -51,7 +47,7 @@ public class JwtUtils {
                     .add(claims)
                     .and()
                 .subject(subject)
-                .issuedAt(java.util.Date.from(Instant.now()))
+                .issuedAt(Date.from(Instant.now()))
                 .expiration(Date.from(Instant.now().plus(jwtExpirationMs, ChronoUnit.MILLIS)))
                 .signWith(privateKey)
                 .compact();
