+++ spring-common-jwt-security/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+ru.t1.nour.security.jwt.config.JwtAutoConfiguration
